+++ frontend/assets/AutoComplete.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace frontend\assets;
+
+use yii\web\AssetBundle;
+
+/**
+ * Main frontend application asset bundle.
+ */
+class AutoComplete extends AssetBundle
+{
+    public $basePath = '@webroot';
+    public $baseUrl = '@web';
+    public $jsOptions = [
+        'position' => \yii\web\View::POS_HEAD,
+    ];
+    public $js = [
+        'https://cdn.jsdelivr.net/npm/@tarekraafat/autocomplete.js@10.2.6/dist/autoComplete.min.js'
+    ];
+}
