+++ frontend/config/params.php
@@ -8,5 +8,6 @@ return [
     'respond' => new Respond(),
     'task' => new Tasks(),
     'review' => new Reviews(),
-    'apikey' => 'e666f398-c983-4bde-8f14-e3fec900592a'
+    'apikey' => 'e666f398-c983-4bde-8f14-e3fec900592a',
+    'uri' => 'https://geocode-maps.yandex.ru/'
 ];
