+++ frontend/controllers/SecuredController.php
@@ -10,7 +10,7 @@ abstract class SecuredController extends Controller
     /**
      * {@inheritdoc}
      */
-    public function behaviors()
+    public function behaviors() : array
     {
         return [
             'access' => [
