+++ frontend/models/Clips.php
@@ -17,7 +17,7 @@ class Clips extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public static function tableName()
+    public static function tableName() : string
     {
         return 'clips';
     }
@@ -25,7 +25,7 @@ class Clips extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public function rules()
+    public function rules() : array
     {
         return [
             [['task_id'], 'integer'],
@@ -38,7 +38,7 @@ class Clips extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public function attributeLabels()
+    public function attributeLabels() : array
     {
         return [
             'id' => 'ID',
