+++ frontend/models/PhotoWork.php
@@ -17,7 +17,7 @@ class PhotoWork extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public static function tableName()
+    public static function tableName() : string
     {
         return 'photo_work';
     }
@@ -25,7 +25,7 @@ class PhotoWork extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public function rules()
+    public function rules() : array
     {
         return [
             [['user_id'], 'integer'],
@@ -39,7 +39,7 @@ class PhotoWork extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public function attributeLabels()
+    public function attributeLabels() : array
     {
         return [
             'id' => 'ID',
