+++ frontend/models/Replies.php
@@ -14,7 +14,7 @@ class Replies extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public static function tableName()
+    public static function tableName() : string
     {
         return 'replies';
     }
@@ -22,7 +22,7 @@ class Replies extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public function rules()
+    public function rules() : array
     {
         return [
             [['date_add', 'rate', 'description'], 'required'],
@@ -35,7 +35,7 @@ class Replies extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public function attributeLabels()
+    public function attributeLabels() : array
     {
         return [
             'id' => 'ID',
