+++ frontend/models/Reviews.php
@@ -20,7 +20,7 @@ class Reviews extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public static function tableName()
+    public static function tableName() : string
     {
         return 'reviews';
     }
@@ -28,7 +28,7 @@ class Reviews extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public function rules()
+    public function rules() : array
     {
         return [
             [['description'], 'string'],
@@ -51,7 +51,7 @@ class Reviews extends \yii\db\ActiveRecord
     /**
      * {@inheritdoc}
      */
-    public function attributeLabels()
+    public function attributeLabels() : array
     {
         return [
             'id' => 'ID',
