+++ frontend/views/mylist/index.php
@@ -9,7 +9,7 @@
 <section class="menu-toggle">
     <ul class="menu-toggle__list">
         <li class="menu-toggle__item
-        <?= \Yii::$app->request->get('status') == 'performed' ? 'menu_toggle__item--current' : '' ?>
+        <?= (string)\Yii::$app->request->get('status') === 'performed' ? 'menu_toggle__item--current' : '' ?>
         menu-toggle__item--completed">
             <div class="menu-toggle__svg-wrapper">
                 <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -21,7 +21,7 @@
                         6.71668 8.85418 6.36239 8.90334C6.00811 8.9525 5.70696 9.18694 5.57239 9.51834C5.43783 9.84974
                         5.49028 10.2278 5.71 10.51L8.15 13.62C8.34082 13.8615 8.63222 14.0017 8.94 14C9.2495 13.9993
                         9.54121 13.8552 9.73 13.61Z"
-                        <?= \Yii::$app->request->get('status') == 'performed' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'performed' ? 'fill="white"' : '' ?> />
                 </svg>
             </div>
             <a href="<?= Url::to(['mylist/index', 'status' => 'performed']) ?>">
@@ -29,7 +29,8 @@
             </a>
         </li>
         <li class="menu-toggle__item
-        <?= \Yii::$app->request->get('status') == 'new' ? 'menu_toggle__item--current' : '' ?> menu-toggle__item--new">
+        <?= (string)\Yii::$app->request->get('status') === 'new' ? 'menu_toggle__item--current' : '' ?>
+        menu-toggle__item--new">
             <div class="menu-toggle__svg-wrapper">
                 <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
                     <rect opacity="0.01" x="24" y="24" width="24" height="24" transform="rotate(180 24 24)"
@@ -37,45 +38,45 @@
                     <path
                         d="M12 6C12.5523 6 13 5.55228 13 5V3C13 2.44772 12.5523 2 12 2C11.4477 2 11 2.44772 11 3V5C11
                         5.55228 11.4477 6 12 6Z"
-                        <?= \Yii::$app->request->get('status') == 'new' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'new' ? 'fill="white"' : '' ?> />
                     <path
                         d="M21 11H19C18.4477 11 18 11.4477 18 12C18 12.5523 18.4477 13 19 13H21C21.5523 13 22 12.5523
                         22 12C22 11.4477 21.5523 11 21 11Z"
-                        <?= \Yii::$app->request->get('status') == 'new' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'new' ? 'fill="white"' : '' ?> />
                     <path
                         d="M6 12C6 11.4477 5.55228 11 5 11H3C2.44772 11 2 11.4477 2 12C2 12.5523 2.44772 13 3
                         13H5C5.55228 13 6 12.5523 6 12Z"
-                        <?= \Yii::$app->request->get('status') == 'new' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'new' ? 'fill="white"' : '' ?> />
                     <path
                         d="M6.22 4.99995C5.81407 4.61611 5.17384 4.63402 4.79 5.03995C4.40616 5.44588 4.42407 6.08611
                         4.83 6.46995L6.27 7.85995C6.46525 8.04848 6.72875 8.14955 7 8.13995C7.27219 8.13891 7.53219
                         8.02696 7.72 7.82995C8.10772 7.43991 8.10772 6.80999 7.72 6.41995L6.22 4.99995Z"
-                        <?= \Yii::$app->request->get('status') == 'new' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'new' ? 'fill="white"' : '' ?> />
                     <path
                         d="M17 8.14C17.2575 8.13898 17.5046 8.03869 17.69 7.86L19.13 6.47C19.4946 6.08908 19.5007
                         5.49053 19.1441 5.10217C18.7874 4.71382 18.1905 4.66911 17.78 5L16.34 6.42C15.9523 6.81004
                         15.9523 7.43997 16.34 7.83C16.5131 8.01272 16.7488 8.12342 17 8.14Z"
-                        <?= \Yii::$app->request->get('status') == 'new' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'new' ? 'fill="white"' : '' ?> />
                     <path
                         d="M12 18C11.4477 18 11 18.4477 11 19V21C11 21.5523 11.4477 22 12 22C12.5523 22 13 21.5523 13
                         21V19C13 18.4477 12.5523 18 12 18Z"
-                        <?= \Yii::$app->request->get('status') == 'new' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'new' ? 'fill="white"' : '' ?> />
                     <path
                         d="M17.73 16.14C17.3324 15.7561 16.6988 15.7673 16.315 16.165C15.9312 16.5626 15.9424 17.1961
                         16.34 17.58L17.78 19C17.9654 19.1786 18.2125 19.2789 18.47 19.28C18.7407 19.2815 19.0005 19.1733
                         19.19 18.98C19.3793 18.7922 19.4858 18.5366 19.4858 18.27C19.4858 18.0033 19.3793 17.7477 19.19
                         17.56L17.73 16.14Z"
-                        <?= \Yii::$app->request->get('status') == 'new' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'new' ? 'fill="white"' : '' ?> />
                     <path
                         d="M6.27 16.14L4.83 17.53C4.64069 17.7178 4.5342 17.9734 4.5342 18.24C4.5342 18.5066 4.64069
                         18.7622 4.83 18.95C5.0195 19.1433 5.27929 19.2516 5.55 19.25C5.79651 19.2521 6.03512 19.1631
                         6.22 19L7.66 17.61C8.05765 17.2262 8.06884 16.5926 7.685 16.195C7.30116 15.7974 6.66765 15.7862
                         6.27 16.17V16.14Z"
-                        <?= \Yii::$app->request->get('status') == 'new' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'new' ? 'fill="white"' : '' ?> />
                     <path
                         d="M12 8C9.79086 8 8 9.79086 8 12C8 14.2091 9.79086 16 12 16C14.2091 16 16 14.2091 16 12C16
                         9.79086 14.2091 8 12 8V8Z"
-                        <?= \Yii::$app->request->get('status') == 'new' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'new' ? 'fill="white"' : '' ?> />
                 </svg>
             </div>
             <a href="<?= Url::to(['mylist/index', 'status' => 'new']) ?>">
@@ -83,7 +84,7 @@
             </a>
         </li>
         <li class="menu-toggle__item
-        <?= \Yii::$app->request->get('status') == 'work' ? 'menu_toggle__item--current' : '' ?>
+        <?= (string)\Yii::$app->request->get('status') === 'work' ? 'menu_toggle__item--current' : '' ?>
         menu-toggle__item--active">
             <div class="menu-toggle__svg-wrapper">
                 <svg width="21" height="19" viewBox="0 0 21 19" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -97,7 +98,7 @@
                         5.55228 16 5 16C4.44771 16 4 15.5523 4 15C4 14.4477 4.44771 14 5 14C5.55228 14 6 14.4477 6
                         15ZM17 16C17.5523 16 18 15.5523 18 15C18 14.4477 17.5523 14 17 14C16.4477 14 16 14.4477 16 15C16
                         15.5523 16.4477 16 17 16Z"
-                        <?= \Yii::$app->request->get('status') == 'work' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'work' ? 'fill="white"' : '' ?> />
                 </svg>
             </div>
             <a href="<?= Url::to(['mylist/index', 'status' => 'work']) ?>">
@@ -105,7 +106,7 @@
             </a>
         </li>
         <li class="menu-toggle__item
-        <?= \Yii::$app->request->get('status') == 'cancel' ? 'menu_toggle__item--current' : '' ?>
+        <?= (string)\Yii::$app->request->get('status') === 'cancel' ? 'menu_toggle__item--current' : '' ?>
         menu-toggle__item--canceled">
             <div class="menu-toggle__svg-wrapper">
                 <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -119,7 +120,7 @@
                         13.0058 8 13.0058C8.26664 13.0058 8.52223 12.8993 8.71 12.71L10 11.41L11.29 12.71C11.4778
                         12.8993 11.7334 13.0058 12 13.0058C12.2666 13.0058 12.5222 12.8993 12.71 12.71C12.8993 12.5222
                         13.0058 12.2666 13.0058 12Z"
-                        <?= \Yii::$app->request->get('status') == 'cancel' ? 'fill="white"' : '' ?> />
+                        <?= (string)\Yii::$app->request->get('status') === 'cancel' ? 'fill="white"' : '' ?> />
                 </svg>
             </div>
             <a href="<?= Url::to(['mylist/index', 'status' => 'cancel']) ?>">
@@ -127,7 +128,7 @@
             </a>
         </li>
         <li class="menu-toggle__item
-        <?= \Yii::$app->request->get('status') == 'failed' ? 'menu_toggle__item--current' : '' ?>
+        <?= (string)\Yii::$app->request->get('status') === 'failed' ? 'menu_toggle__item--current' : '' ?>
         menu-toggle__item--hidden">
             <div class="menu-toggle__svg-wrapper">
                 <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -139,7 +140,7 @@
                             14.6105 14 14.13 14H9.87C9.38951 14 9 13.6105 9 13.13V12.87C9 12.3895 9.38951 12 9.87
                             12H14.13C14.6105 12 15 12.3895 15 12.87V13.13ZM6 7H18C18.5523 7 19 6.55228 19 6C19 5.44772
                             18.5523 5 18 5H6C5.44772 5 5 5.44772 5 6C5 6.55228 5.44772 7 6 7Z"
-                            <?= \Yii::$app->request->get('status') == 'failed' ? 'fill="white"' : '' ?> />
+                            <?= (string)\Yii::$app->request->get('status') === 'failed' ? 'fill="white"' : '' ?> />
                 </svg>
             </div>
             <a href="<?= Url::to(['mylist/index', 'status' => 'failed']) ?>">
